fix 64bit compile issue
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@3977 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ codeview.cpp
@@ -1954,7 +1954,7 @@ bool CodeViewer::ParseOKLineLength(const size_t LineLen)
 		char szText[128];
 		sprintf_s(szText, "The current maximum script line length is %d", MAX_LINE_LENGTH);
 		char szCaption[128];
-		sprintf_s(szCaption, "Line too long on line %d", LineLen);
+		sprintf_s(szCaption, "Line too long on line %d", (int)LineLen);
 		MessageBox(m_hwndMain, szText,szCaption, MB_OK);
 		return false;
 	}
